requirements.txt
asgiref==3.9.1
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
cryptography==46.0.1
defusedxml==0.7.1
dj-database-url==0.5.0
Django==4.2.24
django-allauth==0.57.2
idna==3.10
oauthlib==3.3.1
psycopg2==2.9.10
pycparser==2.23
PyJWT==2.10.1
python3-openid==3.2.0
requests==2.32.5
requests-oauthlib==2.0.0
sqlparse==0.5.3
urllib3==2.5.0
